execution-engine/comm: restructure tests directory
